refactor: migrate coveragePercentage from Double to BigDecimal
Refactor the `coveragePercentage` field in `InsuranceCard` and `Order` models from `Double` to `BigDecimal` to ensure higher precision for financial calculations.

- Update `InsuranceCard` and `Order` entities to use `BigDecimal`.
- Update `PatientMapper`, `AdminService`, `OrderService`, `PatientProfileService`, and `PharmacyMatchingEngine` to handle `BigDecimal` conversions and null safety.
- Update `OrderServiceTest` to use `BigDecimal` for test data.
- Refactor CI/CD workflow to use a dedicated step for lowercasing the image name instead of an environment variable transformation.
- Clean up unused imports and minor formatting in `application.properties`.

+++ src/main/java/com/meddelivery/model/InsuranceCard.java
@@ -5,6 +5,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -41,7 +42,7 @@ public class InsuranceCard {
     private InsuranceStatus status;
 
     @Column(nullable = false, precision = 5, scale = 2)
-    private Double coveragePercentage;
+    private BigDecimal coveragePercentage;
 
     @CreationTimestamp
     private LocalDateTime createdAt;
